Add gulp to build css.
The ability to use the less source for bootstrap is going to be
extremely useful when it comes to customizing things.

Grunt/gulp are nice for making it easier to use, especially with "watch"
behaviors.  I picked gulp to try it out and because it looks really nice
to work with.