Use flexbox to make the footer sticky.
For sparse pages, I don't want the footer in the middle of the page.
Flexbox makes this easy to solve (while gracefully degrading for
non-flexbox browsers to just putting footer in middle of page like it
was before).

In order to make it work well on all the pages, I stuck the main content
inside of a "content" div.  not ideal, but works!  Multiple content
blocks can be added and they will each be given equal flex priority.

The margin-bottom on the blog container was there solely for keeping the
footer down, so that can be removed safely now.